pyproject.toml
[project]
name = "video-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.15"
dependencies = [
    "fastapi>=0.115.6",
    "google-genai>=1.56.0",
    "google-cloud-aiplatform>=1.70.0",
    "inngest>=0.5.0",
    "moviepy>=1.0.3",
    "opencv-python>=4.10.0",
    "faster-whisper>=1.1.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.13.1",
    "streamlit>=1.52.2",
    "uvicorn>=0.32.1",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.3",
]
